app/myauth/serializers.py
from rest_framework import serializers


class GitHubOAuthResponseSerializer(serializers.Serializer):
    access = serializers.CharField()
    refresh = serializers.CharField()
    user = serializers.DictField()
