app/requirements.txt
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
requests
django-cors-headers